+++ src/apytypes_context_wrapper.cc
@@ -1,4 +1,4 @@
-#include "apytypes_common.h"
+#include "apytypes_context.h"
 #include <nanobind/nanobind.h>
 #include <nanobind/stl/function.h>
 #include <nanobind/stl/optional.h>